Bug Fix: repetition of characters were highlighted
suppose the content in the file1 was abbcdeffg and the content of file2 was abb cdeffg. The extra white space and the second f in the two f's were being highlighted. This was because once the first f is matched, its position was added to the matched contents list. This lead to the second f from file 1 to first be matched with the same corresponding position in file1 which is the position of the first f. added a clause that first checks if the position is already there in the matchedIndexes list. If yes, it would try matching it with the successive position in the content from the second file.